refactor(#445): split basic.ts into basicData/colors, letters, shapes, numbers
